Simplify AssemblyScript example, add another
I don't want to mess with npm and `package.json` stuff rn, my solution
is way simpler

+++ langs/assemblyscript/examples/simple/src/env.ts
@@ -10,23 +10,23 @@ export function clearScreen(c: u8): void {
   return _clearScreen(c);
 }
 
-declare function _pSet(x: u8, y: u8, c: u8): void;
-export function pSet(x: u8, y: u8, c: u8): void {
+declare function _pSet(x: i32, y: i32, c: u8): void;
+export function pSet(x: i32, y: i32, c: u8): void {
   _pSet(x, y, c);
 }
 
-declare function _rect(x: u8, y: u8, w: u8, h: u8, c: u8): void;
-export function rect(x: u8, y: u8, w: u8, h: u8, c: u8): void {
+declare function _rect(x: i32, y: i32, w: u32, h: u32, c: u8): void;
+export function rect(x: i32, y: i32, w: u32, h: u32, c: u8): void {
   _rect(x, y, w, h, c);
 }
 
-declare function _rectFill(x: u8, y: u8, w: u8, h: u8, c: u8): void;
-export function rectFill(x: u8, y: u8, w: u8, h: u8, c: u8): void {
+declare function _rectFill(x: i32, y: i32, w: u32, h: u32, c: u8): void;
+export function rectFill(x: i32, y: i32, w: u32, h: u32, c: u8): void {
   _rectFill(x, y, w, h, c);
 }
 
-declare function _sprite(x: u8, y: u8, id: u8): void;
-export function sprite(x: u8, y: u8, id: u8): void {
+declare function _sprite(x: i32, y: i32, id: u8): void;
+export function sprite(x: i32, y: i32, id: u8): void {
   _sprite(x, y, id);
 }
